int Removed and added double

+++ SumOfuserInput.java
@@ -4,10 +4,10 @@ public class SumOfuserInput {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.print("Input the first number: ");
-        int num1 = input.nextInt();
+        Double num1 = input.nextDouble();
         System.out.print("Input the second number: ");
-        int num2 = input.nextInt();
-        int sum = num1 + num2;
+        Double num2 = input.nextDouble();
+        Double sum = num1 + num2;
         System.out.println();
         System.out.println("Sum: " + sum);
     }
